refactor: expand services test support helpers
