+++ crates/luban_backend/src/services/test_support.rs
@@ -1,8 +1,10 @@
 use std::{
-    path::Path,
+    path::{Path, PathBuf},
     process::{Command, Output},
     sync::MutexGuard,
 };
+
+use luban_domain::AttachmentRef;
 
 pub(super) fn lock_env() -> MutexGuard<'static, ()> {
     crate::env::lock_env_for_tests()
@@ -26,4 +28,26 @@ pub(super) fn assert_git_success(repo_path: &Path, args: &[&str]) {
             String::from_utf8_lossy(&output.stderr).trim()
         );
     }
+}
+
+pub(super) fn git_rev_parse(repo_path: &Path, rev: &str) -> String {
+    let out = run_git(repo_path, &["rev-parse", "--verify", rev]);
+    assert!(
+        out.status.success(),
+        "git rev-parse {rev} failed:\nstdout:\n{}\nstderr:\n{}",
+        String::from_utf8_lossy(&out.stdout).trim(),
+        String::from_utf8_lossy(&out.stderr).trim()
+    );
+    String::from_utf8_lossy(&out.stdout).trim().to_owned()
+}
+
+pub(super) fn stored_blob_path(
+    service: &super::GitWorkspaceService,
+    project_slug: &str,
+    workspace_name: &str,
+    attachment: &AttachmentRef,
+) -> PathBuf {
+    service
+        .context_blobs_dir(project_slug, workspace_name)
+        .join(format!("{}.{}", attachment.id, attachment.extension))
 }
